split the IGV into its own component
