+++ src/views/GwasBrowser.js
@@ -1,8 +1,6 @@
-import React, { useRef, useState, useEffect } from "react";
+import React, { useState } from "react";
 import { Row, Input, UncontrolledAlert } from "reactstrap";
-
-// TODO: Importing from igv.esm.min.js is not working, but it works fine in the standalone test
-import igv from 'igv/dist/igv.esm.js'
+import GwasInstance from "../components/IGV/GwasInstance.js"
 
 // Consts
 import BASE_URL from "../constants/constants.js"
@@ -10,28 +8,10 @@ import BASE_URL from "../constants/constants.js"
 
 function GwasBrowser() {
   /***
-   * A functional component that renders an IGV.js instance
+   * A functional component that renders a view with a IGV.js browser.
    */
   const [selectedGwasName, setSelectedGwasName] = useState("");
-  const igvBrowser = useRef(null);
-  const igvOptions = {
-    genome: 'hg38',
-    tracks: [
-      {
-          type: "gwas",
-          format: "gwas",
-          name: "",
-          url: "",
-          indexed: false,
-          height: 300,
-          columns: {
-              chromosome: 1,
-              position: 2,
-              value: 5
-          },
-          dotSize: 2.2
-      }
-  ]};
+  const [selectedGwasUrl, setSelectedGwasUrl] = useState("");
 
   const mock_data = {
     "ANA_A2_V2_filtered": "/static/COVID19_HGI_ANA_A2_V2_20200701.txt.gz_1.0E-5.txt",
@@ -54,20 +34,6 @@ function GwasBrowser() {
       </option>
     );
   });
-
-  useEffect(() => {
-    // Remove an existing browser instance
-    while (igv.getBrowser() !== undefined) {
-      igv.removeBrowser(igv.getBrowser());
-    }
-
-    // Do not create new browser instance on page load as no sample is selected.
-    if (selectedGwasName !== "") {
-      igvOptions["tracks"][0]["name"] = selectedGwasName;
-      igvOptions["tracks"][0]["url"] = BASE_URL + mock_data[selectedGwasName];
-      igv.createBrowser(igvBrowser.current, igvOptions);
-    }
-  }, [selectedGwasName, setSelectedGwasName])
 
   return (
     <>
@@ -94,12 +60,13 @@ function GwasBrowser() {
             <Input defaultValue={'disabled'}
               onChange={(e) => {
                 setSelectedGwasName(e.currentTarget.value);
+                setSelectedGwasUrl(BASE_URL + mock_data[e.currentTarget.value]);
               }}
               type="select">
               { disabledElementList.concat(gwasList) }
             </Input>
 
-            <div className="ml-auto mr-auto" style={{background: "white", marginTop: "15px"}} ref={igvBrowser}></div>
+            <GwasInstance selectedGwasName={selectedGwasName} selectedGwasUrl={selectedGwasUrl} />
         </div>
     </>
   );
